move analyzer test to routing
