+++ eslint-bridge/tests/routing/fixtures/parsing-error.js
@@ -1,0 +1,2 @@
+function f() {
+  return;
