+++ eslint-bridge/tests/routing/routing.test.ts
@@ -22,6 +22,7 @@ import { setContext } from 'helpers';
 import http from 'http';
 import { initializeLinter } from 'linting/eslint';
 import path from 'path';
+import { EMPTY_JSTS_ANALYSIS_OUTPUT } from 'routing/errors';
 import { start } from 'server';
 import { createProgram } from 'services/program';
 import { promisify } from 'util';
@@ -229,13 +230,14 @@ describe('router', () => {
     expect(console.error).toHaveBeenCalled();
   });
 
-  it('should handle errors using the error middleware', async () => {
-    console.error = jest.fn();
-    const tsConfig = path.join(__dirname, 'fixtures', 'malformed.json');
-    const data = { tsConfig };
-    const response = (await request(server, '/analyze-yaml', 'POST', data)) as string;
-    const { error } = JSON.parse(response);
-    expect(error).toBeDefined();
-    expect(console.error).toHaveBeenCalled();
+  it('should return an empty analysis output on parsing errors', async () => {
+    initializeLinter([]);
+
+    const filePath = path.join(__dirname, 'fixtures', 'parsing-error.js');
+    const fileType = 'MAIN';
+    const data = { filePath, fileType, tsConfigs: [] };
+    const response = (await request(server, '/analyze-js', 'POST', data)) as string;
+    const parsedResponse = JSON.parse(response);
+    expect(parsedResponse).toEqual(expect.objectContaining(EMPTY_JSTS_ANALYSIS_OUTPUT));
   });
 });
